BH DCC/DCU: Rename a file
